temp file fixes
