+++ gradio/components.py
@@ -16,7 +16,14 @@ from copy import deepcopy
 from enum import Enum
 from pathlib import Path
 from types import ModuleType
-from typing import TYPE_CHECKING, Any, Callable, Dict, List, Tuple, Type
+from typing import TYPE_CHECKING, Any, Callable, Dict, List, Set, Tuple, Type
+
+import hashlib
+import secrets
+import urllib.request
+import aiofiles
+from fastapi import UploadFile
+import shutil
 
 import altair as alt
 import matplotlib.figure
@@ -28,6 +35,7 @@ from ffmpy import FFmpeg
 from pandas.api.types import is_numeric_dtype
 from PIL import Image as _Image  # using _ to minimize namespace pollution
 from typing_extensions import Literal
+import requests
 
 from gradio import media_data, processing_utils, utils
 from gradio.blocks import Block, BlockContext
@@ -197,6 +205,123 @@ class IOComponent(Component, Serializable):
         )
         if callable(load_fn):
             self.attach_load_event(load_fn, every)
+
+        self.temp_files: Set[str] = set()
+        self.DEFAULT_TEMP_DIR = tempfile.gettempdir()
+
+    def hash_file(self, file_path: str, chunk_num_blocks: int = 128) -> str:
+        sha1 = hashlib.sha1()
+        with open(file_path, "rb") as f:
+            for chunk in iter(lambda: f.read(chunk_num_blocks * sha1.block_size), b""):
+                sha1.update(chunk)
+        return sha1.hexdigest()
+
+    def hash_url(self, url: str, chunk_num_blocks: int = 128) -> str:
+        sha1 = hashlib.sha1()
+        remote = urllib.request.urlopen(url)
+        max_file_size = 100 * 1024 * 1024  # 100MB
+        total_read = 0
+        while True:
+            data = remote.read(chunk_num_blocks * sha1.block_size)
+            total_read += chunk_num_blocks * sha1.block_size
+            if not data or total_read > max_file_size:
+                break
+            sha1.update(data)
+        return sha1.hexdigest()
+
+    def hash_base64(self, base64_encoding: str, chunk_num_blocks: int = 128) -> str:
+        sha1 = hashlib.sha1()
+        for i in range(0, len(base64_encoding), chunk_num_blocks * sha1.block_size):
+            data = base64_encoding[i : i + chunk_num_blocks * sha1.block_size]
+            sha1.update(data.encode("utf-8"))
+        return sha1.hexdigest()
+
+    def make_temp_copy_if_needed(self, file_path: str) -> str:
+        """Returns a temporary file path for a copy of the given file path if it does
+        not already exist. Otherwise returns the path to the existing temp file."""
+        temp_dir = self.hash_file(file_path)
+        temp_dir = Path(self.DEFAULT_TEMP_DIR) / temp_dir
+        temp_dir.mkdir(exist_ok=True, parents=True)
+
+        f = tempfile.NamedTemporaryFile(delete=False, dir=temp_dir)
+        f.name = utils.strip_invalid_filename_characters(Path(file_path).name)
+        full_temp_file_path = str(utils.abspath(temp_dir / f.name))
+
+        if not Path(full_temp_file_path).exists():
+            shutil.copy2(file_path, full_temp_file_path)
+
+        self.temp_files.add(full_temp_file_path)
+        return full_temp_file_path
+
+    async def save_uploaded_file(self, file: UploadFile, upload_dir: str) -> str:
+        temp_dir = secrets.token_hex(
+            20
+        )  # Since the full file is being uploaded anyways, there is no benefit to hashing the file.
+        temp_dir = Path(upload_dir) / temp_dir
+        temp_dir.mkdir(exist_ok=True, parents=True)
+        output_file_obj = tempfile.NamedTemporaryFile(delete=False, dir=temp_dir)
+
+        if file.filename:
+            file_name = Path(file.filename).name
+            output_file_obj.name = utils.strip_invalid_filename_characters(file_name)
+
+        full_temp_file_path = str(utils.abspath(temp_dir / output_file_obj.name))
+
+        async with aiofiles.open(full_temp_file_path, "wb") as output_file:
+            while True:
+                content = await file.read(100 * 1024 * 1024)
+                if not content:
+                    break
+                await output_file.write(content)
+
+        return full_temp_file_path
+
+    def download_temp_copy_if_needed(self, url: str) -> str:
+        """Downloads a file and makes a temporary file path for a copy if does not already
+        exist. Otherwise returns the path to the existing temp file."""
+        temp_dir = self.hash_url(url)
+        temp_dir = Path(self.DEFAULT_TEMP_DIR) / temp_dir
+        temp_dir.mkdir(exist_ok=True, parents=True)
+        f = tempfile.NamedTemporaryFile(delete=False, dir=temp_dir)
+
+        f.name = utils.strip_invalid_filename_characters(Path(url).name)
+        full_temp_file_path = str(utils.abspath(temp_dir / f.name))
+
+        if not Path(full_temp_file_path).exists():
+            with requests.get(url, stream=True) as r:
+                with open(full_temp_file_path, "wb") as f:
+                    shutil.copyfileobj(r.raw, f)
+
+        self.temp_files.add(full_temp_file_path)
+        return full_temp_file_path
+
+    def base64_to_temp_file_if_needed(
+        self, base64_encoding: str, file_name: str | None = None
+    ) -> str:
+        """Converts a base64 encoding to a file and returns the path to the file if
+        the file doesn't already exist. Otherwise returns the path to the existing file."""
+        temp_dir = self.hash_base64(base64_encoding)
+        temp_dir = Path(self.DEFAULT_TEMP_DIR) / temp_dir
+        temp_dir.mkdir(exist_ok=True, parents=True)
+
+        guess_extension = processing_utils.get_extension(base64_encoding)
+        if file_name:
+            file_name = utils.strip_invalid_filename_characters(file_name)
+        elif guess_extension:
+            file_name = "file." + guess_extension
+        else:
+            file_name = "file"
+        f = tempfile.NamedTemporaryFile(delete=False, dir=temp_dir)
+        f.name = file_name
+        full_temp_file_path = str(utils.abspath(temp_dir / f.name))
+
+        if not Path(full_temp_file_path).exists():
+            data, _ = processing_utils.decode_base64_to_binary(base64_encoding)
+            with open(full_temp_file_path, "wb") as fb:
+                fb.write(data)
+
+        self.temp_files.add(full_temp_file_path)
+        return full_temp_file_path
 
     def get_config(self):
         config = {
@@ -1797,7 +1922,6 @@ class Video(
     Uploadable,
     IOComponent,
     FileSerializable,
-    TempFileManager,
 ):
     """
     Creates a video component that can be used to upload/record videos (as an input) or display videos (as an output).
@@ -1854,7 +1978,6 @@ class Video(
         self.include_audio = (
             include_audio if include_audio is not None else source == "upload"
         )
-        TempFileManager.__init__(self)
         IOComponent.__init__(
             self,
             label=label,
@@ -2016,7 +2139,6 @@ class Audio(
     Uploadable,
     IOComponent,
     FileSerializable,
-    TempFileManager,
     TokenInterpretable,
 ):
     """
@@ -2076,7 +2198,6 @@ class Audio(
             raise ValueError(
                 "Audio streaming only available if source is 'microphone'."
             )
-        TempFileManager.__init__(self)
         IOComponent.__init__(
             self,
             label=label,
@@ -2326,7 +2447,6 @@ class File(
     Uploadable,
     IOComponent,
     FileSerializable,
-    TempFileManager,
 ):
     """
     Creates a file component that allows uploading generic file (when used as an input) and or displaying generic files (output).
@@ -2397,7 +2517,6 @@ class File(
         Uses event data gradio.SelectData to carry `value` referring to name of selected file, and `index` to refer to index.
         See EventData documentation on how to use this event data.
         """
-        TempFileManager.__init__(self)
         IOComponent.__init__(
             self,
             label=label,
@@ -3147,7 +3266,6 @@ class UploadButton(
             )
         self.file_types = file_types
         self.label = label
-        TempFileManager.__init__(self)
         IOComponent.__init__(
             self,
             label=label,
@@ -3892,7 +4010,6 @@ class Gallery(IOComponent, TempFileManager, FileSerializable, Selectable):
         Uses event data gradio.SelectData to carry `value` referring to caption of selected image, and `index` to refer to index.
         See EventData documentation on how to use this event data.
         """
-        TempFileManager.__init__(self)
         IOComponent.__init__(
             self,
             label=label,
@@ -4237,7 +4354,6 @@ class Model3D(
             elem_classes: An optional list of strings that are assigned as the classes of this component in the HTML DOM. Can be used for targeting CSS styles.
         """
         self.clear_color = clear_color or [0, 0, 0, 0]
-        TempFileManager.__init__(self)
         IOComponent.__init__(
             self,
             label=label,
